+++ CHANGELOG.md
@@ -4,13 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [0.3.8] - Unreleased
+## [0.3.8] - 2024-03-18
 ### Changed
 - Internal: Added `Upload.get`, `Upload.delete` and `Upload.rename` functions.
 
 ### Fixed
 - Improved value formatting.
 - Unify line endings.
+- UI (draggable): improved drag over hints.
+- UI (object-list, choose-list): open/close on click in the main label.
+- UI (accordion): Make the header 100% width.
 
 ## [0.3.7] - 2024-03-13
 ### Added
@@ -176,7 +179,7 @@ First version
 [#9]: https://github.com/lumeland/cms/issues/9
 [#10]: https://github.com/lumeland/cms/issues/10
 
-[0.3.8]: https://github.com/lumeland/cms/compare/v0.3.7...HEAD
+[0.3.8]: https://github.com/lumeland/cms/compare/v0.3.7...v0.3.8
 [0.3.7]: https://github.com/lumeland/cms/compare/v0.3.6...v0.3.7
 [0.3.6]: https://github.com/lumeland/cms/compare/v0.3.5...v0.3.6
 [0.3.5]: https://github.com/lumeland/cms/compare/v0.3.4...v0.3.5